PR review fixes
