+++ src/Checker/PostUpdateChangesCheckerInterface.php
@@ -4,13 +4,10 @@ declare(strict_types=1);
 
 namespace Setono\SyliusOrderEditPlugin\Checker;
 
-use Setono\SyliusOrderEditPlugin\Entity\InitialTotalAwareOrderInterface;
+use Setono\SyliusOrderEditPlugin\Entity\EditableOrderInterface;
 use Sylius\Component\Core\Model\OrderInterface;
 
 interface PostUpdateChangesCheckerInterface
 {
-    public function check(
-        InitialTotalAwareOrderInterface $previousOrder,
-        OrderInterface $newOrder,
-    ): void;
+    public function check(EditableOrderInterface $previousOrder, OrderInterface $newOrder): void;
 }
