+++ src/Entity/EditableOrderTrait.php
@@ -4,13 +4,29 @@ declare(strict_types=1);
 
 namespace Setono\SyliusOrderEditPlugin\Entity;
 
+use Doctrine\ORM\Mapping as ORM;
+use Sylius\Component\Core\Model\OrderInterface;
 use Sylius\Component\Core\OrderPaymentStates;
 
-/** @method getPaymentState() */
+/** @mixin OrderInterface */
 trait EditableOrderTrait
 {
+    /** @ORM\Column(type="integer") */
+    #[ORM\Column(type: 'integer')]
+    private int $initialTotal = 0;
+
     public function isAlreadyPaid(): bool
     {
         return $this->getPaymentState() === OrderPaymentStates::STATE_PAID;
+    }
+
+    public function getInitialTotal(): int
+    {
+        return $this->initialTotal;
+    }
+
+    public function setInitialTotal(int $initialTotal): void
+    {
+        $this->initialTotal = $initialTotal;
     }
 }
