+++ src/Event/PaidOrderUpdated.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Setono\SyliusOrderEditPlugin\Event;
 
-final class PaidOrderTotalChanged
+final class PaidOrderUpdated
 {
     public function __construct(public int $orderId, public int $oldTotal, public int $newTotal)
     {
