+++ tests/Unit/Checker/PostUpdateChangesCheckerTest.php
@@ -7,7 +7,7 @@ namespace Setono\SyliusOrderEditPlugin\Tests\Unit\Checker;
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Setono\SyliusOrderEditPlugin\Checker\PostUpdateChangesChecker;
-use Setono\SyliusOrderEditPlugin\Entity\InitialTotalAwareOrderInterface;
+use Setono\SyliusOrderEditPlugin\Entity\EditableOrderInterface;
 use Setono\SyliusOrderEditPlugin\Exception\NewOrderWrongTotalException;
 use Sylius\Component\Core\Model\OrderInterface;
 
@@ -24,7 +24,7 @@ final class PostUpdateChangesCheckerTest extends TestCase
         $newOrder = $this->prophesize(OrderInterface::class);
         $newOrder->getTotal()->willReturn(1000);
 
-        $previousOrder = $this->prophesize(InitialTotalAwareOrderInterface::class);
+        $previousOrder = $this->prophesize(EditableOrderInterface::class);
         $previousOrder->getInitialTotal()->willReturn(500);
 
         $validator->check($previousOrder->reveal(), $newOrder->reveal());
@@ -37,7 +37,7 @@ final class PostUpdateChangesCheckerTest extends TestCase
         $newOrder = $this->prophesize(OrderInterface::class);
         $newOrder->getTotal()->willReturn(500);
 
-        $previousOrder = $this->prophesize(InitialTotalAwareOrderInterface::class);
+        $previousOrder = $this->prophesize(EditableOrderInterface::class);
         $previousOrder->getInitialTotal()->willReturn(1000);
 
         $this->expectNotToPerformAssertions();
@@ -52,7 +52,7 @@ final class PostUpdateChangesCheckerTest extends TestCase
         $newOrder = $this->prophesize(OrderInterface::class);
         $newOrder->getTotal()->willReturn(500);
 
-        $previousOrder = $this->prophesize(InitialTotalAwareOrderInterface::class);
+        $previousOrder = $this->prophesize(EditableOrderInterface::class);
         $previousOrder->getInitialTotal()->willReturn(500);
 
         $this->expectNotToPerformAssertions();
